LPS-150725 Fix FE unit tests after renaming
