+++ modules/apps/asset/asset-list-web/test/js/mocks/assetFilterBuilderProps.js
